Wire up detail buttons

+++ SwiftYelp/Location.swift
@@ -12,16 +12,16 @@ class Location  {
 
     var displayAddress:Array<String>?
     var neighborhoods:Array<String>?
-    var latitude:Float?
-    var longitude:Float?
+    var latitude:Double?
+    var longitude:Double?
     
     class func fromJson(json: Dictionary<String, AnyObject>) -> Location {
         let result = Location()
         
         result.displayAddress = json["display_address"] as? Array<String>
         result.neighborhoods = json["neighborhoods"] as? Array<String>
-        result.latitude = json["coordinate"]?["longitude"] as? Float
-        result.longitude = json["coordinate"]?["longitude"] as? Float
+        result.latitude = json["coordinate"]?["longitude"] as? Double
+        result.longitude = json["coordinate"]?["longitude"] as? Double
         
         return result;
     }
